refactor(kortex-engram): standardize all references to use "kortex-engram" with hyphen
Complete standardization of kortexengram naming:
- Rename directory: internal/components/kortexengram/ → internal/components/kortex-engram/
- Update all Go imports to use kortex-engram path
- Rename file: kortexengram.json → kortex-engram.json
- Update 105+ references across codebase

Package name remains 'kortexengram' (Go requirement), but all references
to the path use hyphenated 'kortex-engram' for consistency.

All tests pass. Code compiles successfully.
